Moved under package name and pyproject scaffhold
