+++ ceopardy/exceptions.py
