+++ ceopardy/model.py
