+++ tests/test_utils.py
@@ -7,8 +7,8 @@ and can run without any app context.
 
 import pytest
 
-from exceptions import InvalidQuestionId
-from utils import filter_answer_form, parse_question_id, question_to_html
+from ceopardy.exceptions import InvalidQuestionId
+from ceopardy.utils import filter_answer_form, parse_question_id, question_to_html
 
 # ---------------------------------------------------------------------------
 # parse_question_id
